Move settings->data files
